feat(): implement a utility function creating nestable components
A nestable component is a component that can be nested in itself, in which case the props passed previously will be stacked (except for the children prop). For example, a prop of style defined in one of the ancestor of the rendered element will be applied to the current element as well.

Signed-off-by: David Slutsky <david@evisrael.org>